bugfixes in splrep

+++ src/fpchec.rs
@@ -14,8 +14,8 @@
 ///             t[j] < y[j] < t[j+k+1], j=0,1,2,...,n-k-1
 ///
 ///  The original subroutine in FITPACK by Paul Dierckx is named fpchec
-pub(crate) fn check_knots(x: &Vec<f64>, t: &Vec<f64>, k: u8, m: usize, n: usize) -> u8 {
-    let k1: usize = k as usize + 1;
+pub(crate) fn check_knots(x: &Vec<f64>, t: &Vec<f64>, k: usize, m: usize, n: usize) -> u8 {
+    let k1: usize = k + 1;
     let nk1: usize = n - k1;
     // check condition no 1
     if nk1 < k1 || nk1 > m {
